refactor: split preview page to smaller components
